+++ Devices/lilygo-tdeck/CMakeLists.txt
@@ -1,7 +1,7 @@
-file(GLOB_RECURSE SOURCE_FILES Source/*.c*)
+file(GLOB_RECURSE SOURCE_FILES source/*.c*)
 
 idf_component_register(
     SRCS ${SOURCE_FILES}
-    INCLUDE_DIRS "Source"
+    INCLUDE_DIRS "include"
     REQUIRES Tactility driver
 )
